cleaning and organized the code
